split valid tests into api and cli files; find assertion error
